Extract fulfillment lines as separate component
