+++ src/orders/components/OrderFulfillLine/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./OrderFulfillLine";
+export * from "./OrderFulfillLine";
