+++ src/orders/components/OrderFulfillPage/OrderFulfillPage.tsx
@@ -16,7 +16,6 @@ import PageHeader from "@saleor/components/PageHeader";
 import ResponsiveTable from "@saleor/components/ResponsiveTable";
 import Savebar from "@saleor/components/Savebar";
 import Skeleton from "@saleor/components/Skeleton";
-import TableCellAvatar from "@saleor/components/TableCellAvatar";
 import {
   FulfillOrderMutation,
   OrderFulfillDataQuery,
@@ -26,20 +25,15 @@ import {
 import { SubmitPromise } from "@saleor/hooks/useForm";
 import useFormset, { FormsetData } from "@saleor/hooks/useFormset";
 import { commonMessages } from "@saleor/intl";
-import {
-  Backlink,
-  ConfirmButtonTransitionState,
-  Tooltip,
-  WarningIcon
-} from "@saleor/macaw-ui";
+import { Backlink, ConfirmButtonTransitionState } from "@saleor/macaw-ui";
 import { renderCollection } from "@saleor/misc";
 import { getToFulfillOrderLines } from "@saleor/orders/utils/data";
-import { update } from "@saleor/utils/lists";
 import classNames from "classnames";
 import React from "react";
 import { FormattedMessage, useIntl } from "react-intl";
 
 import { Warehouse } from "../OrderChangeWarehouseDialog/types";
+import OrderFulfillLine from "../OrderFulfillLine/OrderFulfillLine";
 import { messages } from "./messages";
 import { useStyles } from "./styles";
 import { getDefaultFulfillmentValue } from "./utils";
@@ -193,185 +187,16 @@ const OrderFulfillPage: React.FC<OrderFulfillPageProps> = props => {
                         (
                           line: OrderFulfillDataQuery["order"]["lines"][0],
                           lineIndex
-                        ) => {
-                          if (!line) {
-                            return (
-                              <TableRow key={lineIndex}>
-                                <TableCellAvatar className={classes.colName}>
-                                  <Skeleton />
-                                </TableCellAvatar>
-                                <TableCell className={classes.colSku}>
-                                  <Skeleton />
-                                </TableCell>
-                                <TableCell className={classes.colQuantity}>
-                                  <Skeleton />
-                                </TableCell>
-                                <TableCell className={classes.colStock}>
-                                  {" "}
-                                  <Skeleton />
-                                </TableCell>
-                              </TableRow>
-                            );
-                          }
-
-                          const isDeletedVariant = !line?.variant;
-                          const isPreorder = !!line.variant?.preorder;
-                          const lineFormQuantity = isPreorder
-                            ? 0
-                            : formsetData[lineIndex].value?.[0]?.quantity;
-
-                          const overfulfill =
-                            lineFormQuantity > line.quantityToFulfill;
-
-                          const warehouseStock = line.variant?.stocks?.find(
-                            stock => stock.warehouse.id === warehouse.id
-                          );
-
-                          const warehouseAllocation = line.allocations.find(
-                            allocation =>
-                              allocation.warehouse.id === warehouse.id
-                          );
-                          const allocatedQuantityForLine =
-                            warehouseAllocation?.quantity || 0;
-
-                          const availableQuantity =
-                            warehouseStock?.quantity ??
-                            0 - warehouseStock?.quantityAllocated ??
-                            0 + allocatedQuantityForLine;
-
-                          const isStockExceeded =
-                            lineFormQuantity > availableQuantity;
-
-                          return (
-                            <TableRow key={line.id}>
-                              <TableCellAvatar
-                                className={classes.colName}
-                                thumbnail={line?.thumbnail?.url}
-                                badge={
-                                  isPreorder || !line?.variant ? (
-                                    <Tooltip
-                                      variant="warning"
-                                      title={intl.formatMessage(
-                                        isPreorder
-                                          ? messages.preorderWarning
-                                          : messages.deletedVariantWarning
-                                      )}
-                                    >
-                                      <div className={classes.warningIcon}>
-                                        <WarningIcon />
-                                      </div>
-                                    </Tooltip>
-                                  ) : (
-                                    undefined
-                                  )
-                                }
-                              >
-                                {line.productName}
-                                <Typography
-                                  color="textSecondary"
-                                  variant="caption"
-                                >
-                                  {line.variant?.attributes
-                                    ?.map(attribute =>
-                                      attribute.values
-                                        .map(
-                                          attributeValue => attributeValue.name
-                                        )
-                                        .join(", ")
-                                    )
-                                    ?.join(" / ")}
-                                </Typography>
-                              </TableCellAvatar>
-                              <TableCell className={classes.colSku}>
-                                {line.variant?.sku}
-                              </TableCell>
-                              {(() => {
-                                if (isPreorder) {
-                                  return (
-                                    <TableCell
-                                      className={classNames(
-                                        classes.colQuantity,
-                                        classes.error
-                                      )}
-                                    />
-                                  );
-                                }
-
-                                return (
-                                  <TableCell
-                                    className={classes.colQuantity}
-                                    key={
-                                      warehouseStock?.id ??
-                                      "deletedVariant" + lineIndex
-                                    }
-                                  >
-                                    <TextField
-                                      type="number"
-                                      inputProps={{
-                                        className: classNames(
-                                          classes.quantityInnerInput,
-                                          {
-                                            [classes.quantityInnerInputNoRemaining]: !line
-                                              .variant?.trackInventory
-                                          }
-                                        ),
-                                        min: 0,
-                                        style: { textAlign: "right" }
-                                      }}
-                                      fullWidth
-                                      value={lineFormQuantity}
-                                      onChange={event =>
-                                        formsetChange(
-                                          line.id,
-                                          update(
-                                            {
-                                              quantity: parseInt(
-                                                event.target.value,
-                                                10
-                                              ),
-                                              warehouse: warehouse.id
-                                            },
-                                            formsetData[lineIndex].value,
-                                            (a, b) =>
-                                              a.warehouse === b.warehouse
-                                          )
-                                        )
-                                      }
-                                      error={overfulfill}
-                                      variant="outlined"
-                                      InputProps={{
-                                        classes: {
-                                          ...(isStockExceeded &&
-                                            !overfulfill && {
-                                              notchedOutline: classes.warning
-                                            })
-                                        },
-                                        endAdornment: line.variant
-                                          ?.trackInventory && (
-                                          <div
-                                            className={
-                                              classes.remainingQuantity
-                                            }
-                                          >
-                                            / {line.quantityToFulfill}
-                                          </div>
-                                        )
-                                      }}
-                                    />
-                                  </TableCell>
-                                );
-                              })()}
-                              <TableCell
-                                className={classes.colStock}
-                                key="total"
-                              >
-                                {isPreorder || isDeletedVariant
-                                  ? undefined
-                                  : availableQuantity}
-                              </TableCell>
-                            </TableRow>
-                          );
-                        }
+                        ) => (
+                          <OrderFulfillLine
+                            line={line}
+                            lineIndex={lineIndex}
+                            warehouse={warehouse}
+                            formsetData={formsetData}
+                            formsetChange={formsetChange}
+                            formChange={change}
+                          />
+                        )
                       )}
                     </TableBody>
                   </ResponsiveTable>
